refactor: move all ui builder functions to separate files
